+++ src/pages/auth/Signup.jsx
@@ -4,7 +4,7 @@ import Logo from '../../components/common/Logo';
 import { useNavigate } from 'react-router-dom';
 import { useUser } from '../../contexts/UserContext';
 import useAuthRedirect from '../../hooks/useAuthRedirect';
-import SignupForm from '../../components/signup/SignupForm';
+import SignupForm from '../../components/auth/signup/SignupForm';
 
 function Signup() {
     const navigate = useNavigate();
